[mod] mv get_next_line files from gnl to libft
